==> workshops/building-rag-workflows-with-sagemaker-and-bedrock/03-03_raft-customization/05-deploy-to-BR.ipynb <==
# Hosting SageMaker AI models in Bedrock with Bedrock Custom Model Import

In this notebook, you'll take a model artifact that you trained with Amazon SageMaker AI and host it in Amazon Bedrock using Bedrock Custom Model Import.


***

## Global variables

This section contains python variables used in the notebook

import sagemaker
from datasets import load_dataset
import pandas as pd
from transformers import AutoTokenizer
import boto3
import os

sagemaker_session = sagemaker.Session()
bucket_name = sagemaker_session.default_bucket()
account_id = boto3.client('sts').get_caller_identity().get('Account')

print(bucket_name)
print(account_id)

# Import SageMaker AI fine-tuned Model to Amazon Bedrock

This notebook demonstrates how to import models to Amazon Bedrock using Custom Model Import (CMI) feature.

## Prerequisites

- An AWS account with access to Amazon Bedrock
- Appropriate IAM roles and permissions for Bedrock and Amazon S3, following [the instruction here](https://docs.aws.amazon.com/bedrock/latest/userguide/model-import-iam-role.html)
- A S3 bucket prepared to store the custom model
- Sufficient local storage space (At least 17GB for 8B and 135GB for 70B models)


### Step 1: Verify S3 Path

Update these parameters according to your AWS environment:

s3_prefix = f'<<YOUR_FINE_TUNED_MERGED_MODEL>>'

full_s3_path = f"s3://{bucket_name}/{s3_prefix}/"
full_s3_path


!aws s3 ls {full_s3_path}

The Bedrock Custom Model Import job requires a service role to run. The appropriate policies can be found in the [CMI documentation](https://docs.aws.amazon.com/bedrock/latest/userguide/model-import-iam-role.html).

from datetime import datetime

timestamp = datetime.now().strftime("%m-%d-%Y-%H%M%S")
role_name = "bedrock-cmi-role" #replace with your own bedrock service role name if different


# Define your parameters (please update this part based on your setup)
imported_model_name = 'Fine-Tuned-RAFT' # E.x. Deepseek-8B-model
job_name = imported_model_name + '-' + timestamp # E.x. Deepseek-8B-job
role_arn = f'arn:aws:iam::{account_id}:role/{role_name}' # Please make sure it has sufficient permission as listed in the pre-requisite

# Region (currently only 'us-west-2' and 'us-east-1' support CMI with Deepseek-Distilled-Llama models)
region_info = sagemaker_session.boto_region_name#'us-west-2' # You can modify to 'us-east-1' based on your need

job_name

### Step 2: Create Custom Model Import Job

Initialize the import job in Amazon Bedrock

<div class="alert alert-warning">
<b>Note:</b> Creating CMI job for 8B model could take 5-20 minutes to complete.
</div>

# Initialize the Bedrock client
bedrock = boto3.client('bedrock', region_name=region_info)

s3_uri = f's3://{bucket_name}/{s3_prefix}/'

# Create the model import job
response = bedrock.create_model_import_job(
    jobName=job_name,
    importedModelName=imported_model_name,
    roleArn=role_arn,
    modelDataSource={
        's3DataSource': {
            's3Uri': s3_uri
        }
    }
)

job_Arn = response['jobArn']

# Output the job ARN
print(f"Model import job created with ARN: {response['jobArn']}")


### Step 3: Monitor Import Job Status

Check the status of your import job

import time
# Check CMI job status
while True:
    response = bedrock.get_model_import_job(jobIdentifier=job_Arn)
    status = response['status'].upper()
    print(f"Status: {status}")
    
    if status in ['COMPLETED', 'FAILED']:
        break
        
    time.sleep(60)  # Check every 60 seconds

# Get the model ID
model_id = response['importedModelArn']

### Step 4: Wait for Model Initialization

Allow time for the model to initialize:

model_id

# Wait for 5mins for cold start 
time.sleep(300)

### Step 5: Model Inference

from botocore.config import Config
import json

def format_messages(messages: list[dict[str, str]]) -> list[str]:
    """
    Format messages for Llama 3+ chat models.
    
    The model only supports 'system', 'user' and 'assistant' roles, starting with 'system', then 'user' and 
    alternating (u/a/u/a/u...). The last message must be from 'user'.
    """
    # auto assistant suffix
    # messages.append({"role": "assistant"})
    
    output = "<|begin_of_text|>"
    # Adding an inferred prefix
    system_prefix = f"\n\nCutting Knowledge Date: December 2024\nToday Date: {datetime.now().strftime('%d %b %Y')}\n\n"
    for i, entry in enumerate(messages):
        output += f"<|start_header_id|>{entry['role']}<|end_header_id|>"
        if entry['role'] == 'system':
            output += f"{system_prefix}{entry['content']}<|eot_id|>"
        elif entry['role'] != 'system' and 'content' in entry:
            output += f"\n\n{entry['content']}<|eot_id|>"
    output += "<|start_header_id|>assistant<|end_header_id|>\n"
    return output


def send_prompt(messages, temperature=0.3, max_tokens=4096, top_p=0.9, continuation=False, max_retries=10):
    # convert u/a format 
    frmt_input = format_messages(messages)

    client = boto3.Session().client(
    service_name='bedrock-runtime',
    region_name=region_info,
    config=Config(
        connect_timeout=300,  # 5 minutes
        read_timeout=300,     # 5 minutes
        retries={'max_attempts': 3}
        )
    )

    attempt = 0
    while attempt < max_retries:
        try:
            response = client.invoke_model(
                modelId=model_id,
                body=json.dumps({
                    'prompt': frmt_input,
                    'temperature': temperature,
                    'max_gen_len': max_tokens,
                    'top_p': top_p
                }),
                accept='application/json',
                contentType='application/json'
            )
            
            result = json.loads(response['body'].read().decode('utf-8'))
            return result
            
        except Exception as e:
            print(f"Attempt {attempt + 1} failed: {str(e)}")
            attempt += 1
            if attempt < max_retries:
                time.sleep(30)
    
    raise Exception("Failed to get response after maximum retries")
    
    return response

def build_messages(data):
    system_content = f"""You are an assistant for question-answering tasks. Answer the following question in 5 sentences using the provided context. If you don't know the answer, just say "I don't know."."""
    user_content = f"""
        Context: {data["CONTEXT"]} 
        
        Question: {data["QUESTION"]}
        """

    messages = [
        {"role": "system", "content": system_content},
        {"role": "user", "content": user_content}
    ]
    
    return messages

from datasets import load_dataset, concatenate_datasets

test_dataset = load_dataset("json", data_files="../data/sft_test_data.json", split="train")

test_item = test_dataset[2]
test_item

messages = build_messages(test_item)

model_response = send_prompt(messages)

print(f"""
    ============== Question ============
    {test_item["QUESTION"]}
    
    ========= Generated Answer =========
    {model_response}

    ======== Ground Truth Answer =======
    {test_item["ANSWER"]}
    """
)
